@@ -1,6 +1,7 @@
 {
   "name": "app",
   "version": "0.1.0",
+  "type": "module",
   "private": true,
   "engines": {
     "node": "20.19.5",
@@ -23,13 +24,13 @@
     "ts:check": "tsc --noEmit"
   },
   "dependencies": {
-    "@trussworks/react-uswds": "^9.0.0",
-    "@uswds/uswds": "3.8.1",
+    "@trussworks/react-uswds": "^10.0.2",
+    "@uswds/uswds": "^3.13.0",
     "lodash": "^4.17.21",
-    "next": "^15.1.4",
+    "next": "^15.5.2",
     "next-intl": "^3.2.1",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
+    "react": "^19.1.1",
+    "react-dom": "^19.1.1",
     "server-only": "^0.0.1"
   },
   "devDependencies": {
@@ -38,29 +39,29 @@
     "@storybook/nextjs": "^8.0.0",
     "@storybook/react": "^8.0.0",
     "@testing-library/dom": "^10.0.0",
-    "@testing-library/jest-dom": "^6.0.0",
+    "@testing-library/jest-dom": "^6.8.0",
     "@testing-library/react": "^16.0.0",
     "@testing-library/user-event": "^14.4.3",
-    "@types/jest": "^29.5.5",
-    "@types/jest-axe": "^3.5.5",
+    "@types/jest": "^30.0.0",
+    "@types/jest-axe": "^3.5.9",
     "@types/lodash": "^4.14.202",
     "@types/node": "^22.10.2",
-    "@types/react": "^18.2.42",
-    "@types/react-dom": "^18.2.17",
+    "@types/react": "^19.1.12",
+    "@types/react-dom": "^19.1.9",
     "@typescript-eslint/eslint-plugin": "^8.18.1",
     "@typescript-eslint/parser": "^8.18.1",
     "eslint": "^9.0.0",
-    "eslint-config-next": "^15.1.1",
+    "eslint-config-next": "^15.5.2",
     "eslint-config-prettier": "^9.0.0",
-    "eslint-plugin-jest": "^28.0.0",
-    "eslint-plugin-jest-dom": "^5.0.1",
+    "eslint-plugin-jest": "^29.0.1",
+    "eslint-plugin-jest-dom": "^5.5.0",
     "eslint-plugin-storybook": "^0.12.0",
     "eslint-plugin-testing-library": "^7.0.0",
     "eslint-plugin-you-dont-need-lodash-underscore": "^6.13.0",
-    "jest": "^29.5.0",
-    "jest-axe": "^9.0.0",
-    "jest-cli": "^29.5.0",
-    "jest-environment-jsdom": "^29.5.0",
+    "jest": "^30.1.3",
+    "jest-axe": "^10.0.0",
+    "jest-cli": "^30.1.3",
+    "jest-environment-jsdom": "^30.1.2",
     "postcss": "^8.4.31",
     "postcss-loader": "^8.0.0",
     "postcss-preset-env": "^10.0.0",
@@ -69,6 +70,7 @@
     "sass-loader": "^16.0.0",
     "storybook": "^8.6.14",
     "style-loader": "^4.0.0",
+    "ts-node": "^10.9.2",
     "typescript": "^5.0.0"
   }
 }
