@@ -1,7 +1,7 @@
 // @ts-check
 
 /** @type {import("@ianvs/prettier-plugin-sort-imports").PrettierConfig} */
-module.exports = {
+export default {
   /**
    * Sort imports so that the order is roughly:
    * - Built-in Node modules (`import ... from "fs"`)
@@ -11,6 +11,7 @@ module.exports = {
    * - All other files
    * @see https://github.com/IanVS/prettier-plugin-sort-imports
    */
+  plugins: ["@ianvs/prettier-plugin-sort-imports"],
   importOrder: [
     "<BUILTIN_MODULES>",
     "<THIRD_PARTY_MODULES>",
