@@ -3,17 +3,13 @@
  * Handles things like config for location of story files and managing presets (which configure webpack and babel).
  * @see https://storybook.js.org/docs/configurations/default-config/
  */
-// @ts-check
-
-import path from "path";
+import * as path from "path";
+import type { StorybookConfig } from "@storybook/nextjs";
 
 // Support deploying to a subdirectory, such as GitHub Pages.
 const NEXT_PUBLIC_BASE_PATH = process.env.NEXT_PUBLIC_BASE_PATH ?? "";
 
-/**
- * @type {import("@storybook/nextjs").StorybookConfig}
- */
-const config = {
+const config: StorybookConfig = {
   stories: [
     "../stories/**/*.@(mdx|stories.@(js|jsx|ts|tsx))",
     "../src/**/*.@(mdx|stories.@(js|jsx|ts|tsx))",
@@ -26,7 +22,11 @@ const config = {
     // https://storybook.js.org/docs/get-started/nextjs
     name: "@storybook/nextjs",
     options: {
-      nextConfigPath: path.resolve(import.meta.dirname, "../next.config.js"),
+      // Storybook uses esbuild to resolve TypeScript files targeting the CJS
+      // module system, so use `__dirname` instead of `import.meta.dirname`.
+      //
+      // https://github.com/storybookjs/storybook/blob/50c5b26b67bea71442b59e03cbba718c99acaf8e/code/core/src/common/utils/interpret-require.ts#L13
+      nextConfigPath: path.resolve(__dirname, "../next.config.ts"),
       builder: {
         // lazyCompilation breaks Storybook when running from within Docker
         // Google Translate this page for context: https://zenn.dev/yutaosawa/scraps/7764e5f17173d1
